+++ python/lsst/ts/mtdome/mock_llc/base_mock_llc.py
@@ -19,17 +19,31 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-"""Sphinx configuration file for an LSST stack package.
+__all__ = ["BaseMockStatus"]
 
-This configuration only affects single-package Sphinx documentation builds.
-"""
+from abc import ABC, abstractmethod
+from typing import Any, Dict
 
-from documenteer.conf.pipelinespkg import *  # noqa
-import lsst.ts.mtdome  # noqa
 
-project = "ts_mtdome"
-html_theme_options["logotext"] = project  # type: ignore # noqa
-html_title = project
-html_short_title = project
+class BaseMockStatus(ABC):
+    """Abstract base class for all mock status classes used by the mock
+    controller when in simulator mode.
+    """
 
-intersphinx_mapping["ts_xml"] = ("https://ts-xml.lsst.io", None)  # type: ignore # noqa
+    def __init__(self) -> None:
+        # dict to hold the status of the Lower Level Component.
+        self.llc_status: Dict[str, Any] = {}
+        # time of the last executed command, in TAI Unix seconds
+        self.command_time_tai = 0
+
+    @abstractmethod
+    async def determine_status(self, current_tai: float) -> None:
+        """Abstract method that determines the status of the Lower Level
+        Component to be implemented by all concrete sub-classes.
+
+        Parameters
+        ----------
+        current_tai: `float`
+            The current Unix TAI time, in seconds
+        """
+        pass