+++ tests/test_azimuth_motion.py
@@ -1,4 +1,4 @@
-# This file is part of ts_MTDome.
+# This file is part of ts_mtdome.
 #
 # Developed for the Vera Rubin Observatory Telescope and Site Systems.
 # This product includes software developed by the LSST Project
@@ -23,7 +23,7 @@ import logging
 import math
 import unittest
 
-from lsst.ts import MTDome
+from lsst.ts import mtdome
 
 logging.basicConfig(
     format="%(asctime)s:%(levelname)s:%(name)s:%(message)s", level=logging.DEBUG
@@ -48,7 +48,7 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
         start_tai: `float`
             The start TAI time.
         """
-        self.azimuth_motion = MTDome.mock_llc.AzimuthMotion(
+        self.azimuth_motion = mtdome.mock_llc.AzimuthMotion(
             start_position=start_position,
             max_speed=max_speed,
             start_tai=start_tai,
@@ -60,7 +60,7 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
         target_position: float,
         crawl_velocity: float,
         expected_duration: float,
-        motion_state: MTDome.LlcMotionState,
+        motion_state: mtdome.LlcMotionState,
     ) -> None:
         """Verify that the AzimuthMotion computes the correct
         duration.
@@ -75,8 +75,8 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             The crawl velocity after the motion.
         expected_duration: `float`
             The expected duration.
-        motion_state: `MTDome.LlcMotionState`
-            The commanded MTDome.LlcMotionState.
+        motion_state: `mtdome.LlcMotionState`
+            The commanded mtdome.LlcMotionState.
         """
         duration = self.azimuth_motion.set_target_position_and_velocity(
             start_tai=start_tai,
@@ -91,7 +91,7 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
         tai: float,
         expected_position: float,
         expected_velocity: float,
-        expected_motion_state: MTDome.LlcMotionState,
+        expected_motion_state: mtdome.LlcMotionState,
     ) -> None:
         """Verify the position of the AzimuthMotion at the given TAI
         time.
@@ -136,31 +136,31 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(4.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(8.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.5,
             expected_position=math.radians(10.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 4.0,
             expected_position=math.radians(10.15),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_move_zero_ten_neg(self) -> None:
@@ -183,31 +183,31 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(4.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(8.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.5,
             expected_position=math.radians(10.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 4.0,
             expected_position=math.radians(9.85),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_move_ten_zero_pos(self) -> None:
@@ -230,31 +230,31 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(6.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(2.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.5,
             expected_position=math.radians(0.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 4.0,
             expected_position=math.radians(0.15),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_move_ten_zero_neg(self) -> None:
@@ -277,31 +277,31 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(6.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(2.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.5,
             expected_position=math.radians(0.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 4.0,
             expected_position=math.radians(359.85),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_move_ten_threefifty_pos(self) -> None:
@@ -326,37 +326,37 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(6.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(2.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(358.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(350.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 6.0,
             expected_position=math.radians(350.1),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_move_ten_threefifty_neg(self) -> None:
@@ -381,37 +381,37 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(6.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(2.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(358.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(350.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 6.0,
             expected_position=math.radians(349.9),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_move_threefifty_ten_pos(self) -> None:
@@ -436,37 +436,37 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(354.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(358.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(2.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(10.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 6.0,
             expected_position=math.radians(10.1),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_move_threefifty_ten_neg(self) -> None:
@@ -491,37 +491,37 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(354.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(358.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(2.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(10.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 6.0,
             expected_position=math.radians(9.9),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_crawl_pos(self) -> None:
@@ -545,43 +545,43 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.CRAWLING,
+            motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(351.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(352.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 10.0,
             expected_position=math.radians(0.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 11.0,
             expected_position=math.radians(1.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 20.0,
             expected_position=math.radians(10.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 21.0,
             expected_position=math.radians(11.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_crawl_neg(self) -> None:
@@ -605,43 +605,43 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.CRAWLING,
+            motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(9.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 2.0,
             expected_position=math.radians(8.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 10.0,
             expected_position=math.radians(0.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 11.0,
             expected_position=math.radians(359.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 20.0,
             expected_position=math.radians(350.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 21.0,
             expected_position=math.radians(349.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
 
     async def test_stop_from_moving(self) -> None:
@@ -664,20 +664,20 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(4.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         self.azimuth_motion.stop(start_tai=_start_tai + 2.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(8.0),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.STOPPED,
+            expected_motion_state=mtdome.LlcMotionState.STOPPED,
         )
 
     async def test_stop_from_crawling_after_moving(self) -> None:
@@ -700,26 +700,26 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(4.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(10.05),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         self.azimuth_motion.stop(start_tai=_start_tai + 4.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(10.15),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.STOPPED,
+            expected_motion_state=mtdome.LlcMotionState.STOPPED,
         )
 
     async def test_stop_from_crawling(self) -> None:
@@ -742,20 +742,20 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.CRAWLING,
+            motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(11.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         self.azimuth_motion.stop(start_tai=_start_tai + 4.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(14.0),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.STOPPED,
+            expected_motion_state=mtdome.LlcMotionState.STOPPED,
         )
 
     async def test_park_from_moving(self) -> None:
@@ -778,26 +778,26 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(4.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         self.azimuth_motion.park(start_tai=_start_tai + 2.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(4.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.PARKING,
+            expected_motion_state=mtdome.LlcMotionState.PARKING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 4.0,
             expected_position=math.radians(0.0),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.PARKED,
+            expected_motion_state=mtdome.LlcMotionState.PARKED,
         )
 
     async def test_park_from_crawling_after_moving(self) -> None:
@@ -820,38 +820,38 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(4.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(10.05),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         self.azimuth_motion.park(start_tai=_start_tai + 4.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(6.15),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.PARKING,
+            expected_motion_state=mtdome.LlcMotionState.PARKING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 6.0,
             expected_position=math.radians(2.15),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.PARKING,
+            expected_motion_state=mtdome.LlcMotionState.PARKING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 7.0,
             expected_position=math.radians(0.0),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.PARKED,
+            expected_motion_state=mtdome.LlcMotionState.PARKED,
         )
 
     async def test_park_from_crawling(self) -> None:
@@ -874,38 +874,38 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.CRAWLING,
+            motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(11.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         self.azimuth_motion.park(start_tai=_start_tai + 4.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(10.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.PARKING,
+            expected_motion_state=mtdome.LlcMotionState.PARKING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 6.0,
             expected_position=math.radians(6.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.PARKING,
+            expected_motion_state=mtdome.LlcMotionState.PARKING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 7.0,
             expected_position=math.radians(2.0),
             expected_velocity=-_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.PARKING,
+            expected_motion_state=mtdome.LlcMotionState.PARKING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 8.0,
             expected_position=math.radians(0.0),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.PARKED,
+            expected_motion_state=mtdome.LlcMotionState.PARKED,
         )
 
     async def test_stationary_from_moving(self) -> None:
@@ -928,20 +928,20 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(4.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         self.azimuth_motion.go_stationary(start_tai=_start_tai + 2.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(8.0),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.STATIONARY,
+            expected_motion_state=mtdome.LlcMotionState.STATIONARY,
         )
 
     async def test_stationary_from_crawling_after_moving(self) -> None:
@@ -964,26 +964,26 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.MOVING,
+            motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(4.0),
             expected_velocity=_MAX_SPEED,
-            expected_motion_state=MTDome.LlcMotionState.MOVING,
+            expected_motion_state=mtdome.LlcMotionState.MOVING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 3.0,
             expected_position=math.radians(10.05),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         self.azimuth_motion.go_stationary(start_tai=_start_tai + 4.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(10.15),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.STATIONARY,
+            expected_motion_state=mtdome.LlcMotionState.STATIONARY,
         )
 
     async def test_stationary_from_crawling(self) -> None:
@@ -1006,20 +1006,20 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
             target_position=target_position,
             crawl_velocity=crawl_velocity,
             expected_duration=expected_duration,
-            motion_state=MTDome.LlcMotionState.CRAWLING,
+            motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         await self.verify_azimuth_motion(
             tai=_start_tai + 1.0,
             expected_position=math.radians(11.0),
             expected_velocity=crawl_velocity,
-            expected_motion_state=MTDome.LlcMotionState.CRAWLING,
+            expected_motion_state=mtdome.LlcMotionState.CRAWLING,
         )
         self.azimuth_motion.go_stationary(start_tai=_start_tai + 4.0)
         await self.verify_azimuth_motion(
             tai=_start_tai + 5.0,
             expected_position=math.radians(14.0),
             expected_velocity=0,
-            expected_motion_state=MTDome.LlcMotionState.STATIONARY,
+            expected_motion_state=mtdome.LlcMotionState.STATIONARY,
         )
 
     async def test_too_high(self) -> None:
@@ -1043,7 +1043,7 @@ class AzimuthMotionTestCase(unittest.IsolatedAsyncioTestCase):
                 target_position=target_position,
                 crawl_velocity=crawl_velocity,
                 expected_duration=expected_duration,
-                motion_state=MTDome.LlcMotionState.MOVING,
+                motion_state=mtdome.LlcMotionState.MOVING,
             )
             self.fail("Expected a ValueError.")
         except ValueError: